Removes on_message response.
It's just, not feasible to remove the mention
at the beginning of the message to the bot.
At least, it's not feasible to do it succinctly.
A much easier way is to just talk to the bot through
a command.
I opted to create a "speak" command which is basically
an alias for "speak_to_alice", but without acknowledging
that it uses Alice as the backend.